+++ src/Faker/Music.hs
@@ -3,9 +3,8 @@
 
 module Faker.Music where
 
-import Data.Text
-import Faker
-import Faker.Internal
+import Data.Text (Text)
+import Faker (Fake(..))
 import Faker.Provider.Music
 import Faker.TH
 
@@ -19,3 +18,13 @@ $(generateFakeField "music" "genres")
 
 -- | @since 1.0
 $(generateFakeField "music" "mambo_no_5")
+
+
+
+$(generateFakeFields "music" ["hiphop","subgenres"])
+
+$(generateFakeFields "music" ["hiphop","groups"])
+
+$(generateFakeFields "music" ["hiphop","artist"])
+
+