+++ test/BirdSpec.hs
@@ -8,7 +8,7 @@ import Data.Text hiding (all, map)
 import qualified Data.Text as T
 import qualified Data.Vector as V
 import Faker hiding (defaultFakerSettings)
-import Faker.Bird
+import Faker.Creature.Bird
 import Test.Hspec
 import TestImport
 import Faker.Internal